Update to new microcode storage method
** BREAKS BACKWARDS COMPATIBILITY **

Instead of exporting the microcode image as a constant array and length, compile it statically into libdiscferret and add a function 'discferret_fpga_load_default()' which loads the default microcode image into the FPGA.

This fixes the issue where applications would break if libdiscferret's microcode image changed size. Applications will have to be updated to suit this new API, but this is a one-line change.